+++ Reminder/PetReminder.cs
@@ -34,7 +34,7 @@ namespace RemindMe.Reminder {
             return true;
         }
 
-        public override ushort GetIconID(DalamudPluginInterface pluginInterface, RemindMe plugin, MonitorDisplay display) {
+        public override uint GetIconID(DalamudPluginInterface pluginInterface, RemindMe plugin, MonitorDisplay display) {
             try {
                 var action = Service.Data.Excel.GetSheet<Action>().GetRow(petIcons[Service.ClientState.LocalPlayer.ClassJob.Id]);
                 return action.Icon;
